refactor(server): Utilize "usize" for stream ID and topic ID (#2246)
## Summary

According to #2198, align ID types of stream and topic with the target
pointer size "usize" to avoid churny casts and platform-specific
inconsistencies.

Some parts outside the scope of server needs to be changed due to the
depency for structs under iggy common.

+++ core/server/src/shard/system/messages.rs
@@ -61,8 +61,8 @@ impl IggyShard {
             .borrow()
             .append_messages(
                 user_id,
-                numeric_stream_id as u32,
-                numeric_topic_id as u32,
+                numeric_stream_id,
+                numeric_topic_id,
             )
             .with_error_context(|error| {
                 format!("{COMPONENT} (error: {error}) - permission denied to append messages for user {} on stream ID: {}, topic ID: {}", user_id, numeric_stream_id as u32, numeric_topic_id as u32)
@@ -170,7 +170,7 @@ impl IggyShard {
 
         self.permissioner
             .borrow()
-            .poll_messages(user_id, numeric_stream_id as u32, numeric_topic_id as u32)
+            .poll_messages(user_id, numeric_stream_id, numeric_topic_id)
             .with_error_context(|error| format!(
                 "{COMPONENT} (error: {error}) - permission denied to poll messages for user {} on stream ID: {}, topic ID: {}",
                 user_id,
@@ -363,8 +363,8 @@ impl IggyShard {
             .borrow()
             .append_messages(
                 session.get_user_id(),
-                numeric_stream_id as u32,
-                numeric_topic_id as u32,
+                numeric_stream_id,
+                numeric_topic_id,
             )
             .with_error_context(|error| {
                 format!("{COMPONENT} (error: {error}) - permission denied to append messages for user {} on stream ID: {}, topic ID: {}", session.get_user_id(), numeric_stream_id as u32, numeric_topic_id as u32)
